feat(menu): fix mocks in menu tests
