+++ packages/paste-core/components/menu/__tests__/customization.spec.tsx
@@ -5,20 +5,10 @@ import type {RenderOptions} from '@testing-library/react';
 import {CustomizationProvider} from '@twilio-paste/customization';
 
 import {initStyles, BaseMenu} from '../stories/customization.stories';
-import {mockUseMenuStateReturn} from './mocks';
 
 expect.extend(matchers);
 
 const PASTE_ELEMENT = 'data-paste-element';
-
-jest.mock('../src', () => {
-  const actual = jest.requireActual('../src');
-
-  return {
-    ...actual,
-    useMenuState: ({baseId}: {baseId: string}) => mockUseMenuStateReturn(baseId),
-  };
-});
 
 const initWrapper = (element: string): RenderOptions['wrapper'] => ({children}) => (
   <CustomizationProvider
@@ -396,4 +386,79 @@ describe('Menu Customization', () => {
       });
     });
   });
+});
+
+const NOOP = (): null => null;
+
+jest.mock('../src/useMenuState', () => {
+  return {
+    useMenuState: (baseId: string): Record<string, any> => ({
+      animated: false,
+      animating: false,
+      baseId,
+      currentId: null,
+      down: NOOP,
+      first: NOOP,
+      groups: [],
+      hide: NOOP,
+      items: [
+        {disabled: false, groupId: undefined, id: 'primary-menu-customization-story-1', ref: {current: 'button'}},
+        {disabled: false, groupId: undefined, id: 'primary-menu-customization-story-2', ref: {}},
+        {disabled: false, groupId: undefined, id: 'primary-menu-customization-story-3', ref: {}},
+        {disabled: false, groupId: undefined, id: 'primary-menu-customization-story-4', ref: {}},
+        {disabled: false, groupId: undefined, id: 'primary-menu-customization-story-5', ref: {}},
+        {disabled: false, groupId: undefined, id: 'primary-menu-customization-story-6', ref: {}},
+      ],
+      last: NOOP,
+      loop: false,
+      modal: false,
+      move: NOOP,
+      next: NOOP,
+      orientation: 'vertical',
+      place: NOOP,
+      placement: 'bottom-start',
+      previous: NOOP,
+      registerGroup: NOOP,
+      registerItem: NOOP,
+      reset: NOOP,
+      rtl: false,
+      setAnimated: NOOP,
+      setBaseId: NOOP,
+      setCurrentId: NOOP,
+      setLoop: NOOP,
+      setModal: NOOP,
+      setOrientation: NOOP,
+      setRTL: NOOP,
+      setShift: NOOP,
+      setVisible: NOOP,
+      setWrap: NOOP,
+      shift: false,
+      show: NOOP,
+      sort: NOOP,
+      stopAnimation: NOOP,
+      toggle: NOOP,
+      unregisterGroup: NOOP,
+      unregisterItem: NOOP,
+      unstable_arrowRef: {current: null},
+      unstable_arrowStyles: {},
+      unstable_disclosureRef: {current: undefined},
+      unstable_hasActiveWidget: false,
+      unstable_idCountRef: {current: 6},
+      unstable_includesBaseElement: false,
+      unstable_moves: 2,
+      unstable_originalPlacement: 'bottom-start',
+      unstable_popoverRef: {current: undefined},
+      unstable_referenceRef: {current: undefined},
+      unstable_setHasActiveWidget: NOOP,
+      unstable_setIncludesBaseElement: NOOP,
+      unstable_setValue: NOOP,
+      unstable_setVirtual: NOOP,
+      unstable_update: NOOP,
+      unstable_values: {},
+      unstable_virtual: false,
+      up: NOOP,
+      visible: true,
+      wrap: false,
+    }),
+  };
 });
