+++ packages/paste-core/components/menu/src/index.tsx
@@ -1,6 +1,4 @@
-import {useMenuPrimitiveState} from '@twilio-paste/menu-primitive';
-
-export {useMenuPrimitiveState as useMenuState};
+export * from './useMenuState';
 export * from './Menu';
 export * from './MenuButton';
 export * from './MenuGroup';
