+++ packages/paste-core/components/menu/src/useMenuState.ts
@@ -1,0 +1,3 @@
+import {useMenuPrimitiveState} from '@twilio-paste/menu-primitive';
+
+export {useMenuPrimitiveState as useMenuState};
